refactor: merge use-remove-data, use-stake-data, use-unstake-data into use-positions-rewards
