+++ packages/web/src/components/unstake/select-unstake-result/SelectUnstakeResult.tsx
@@ -1,10 +1,12 @@
-import { useUnstakeData } from "@hooks/stake/use-unstake-data";
-import { PoolPositionModel } from "@models/position/pool-position-model";
 import React from "react";
-import { wrapper } from "./SelectUnstakeResult.styles";
-import MissingLogo from "@components/common/missing-logo/MissingLogo";
-import { formatPoolPairAmount } from "@utils/new-number-utils";
 import { useTranslation } from "react-i18next";
+
+import MissingLogo from "@components/common/missing-logo/MissingLogo";
+import { usePositionsRewards } from "@hooks/position/use-positions-rewards";
+import { PoolPositionModel } from "@models/position/pool-position-model";
+import { formatPoolPairAmount } from "@utils/new-number-utils";
+
+import { wrapper } from "./SelectUnstakeResult.styles";
 
 interface SelectUnstakeResultProps {
   positions: PoolPositionModel[];
@@ -15,7 +17,7 @@ const SelectUnstakeResult: React.FC<SelectUnstakeResultProps> = ({
 }) => {
   const { t } = useTranslation();
   const { pooledTokenInfos, unclaimedRewards, totalLiquidityUSD } =
-    useUnstakeData({ positions });
+    usePositionsRewards({ positions });
 
   if (positions.length === 0) return <></>;
   return (
@@ -42,27 +44,28 @@ const SelectUnstakeResult: React.FC<SelectUnstakeResultProps> = ({
         {unclaimedRewards.map((pooledTokenInfo, index) => {
           if (!pooledTokenInfo.amount) return null;
           return (
-          <li key={index}>
-            <div className="main-info">
-              <MissingLogo
-                symbol={pooledTokenInfo.token.symbol}
-                url={pooledTokenInfo.token.logoURI}
-                width={24}
-                mobileWidth={24}
-              />
-              <p>
-                {t("UnstakePosition:overview.unclaimed")}{" "}
-                {pooledTokenInfo.token.symbol}
-              </p>
-              <strong>
-                {formatPoolPairAmount(pooledTokenInfo.amount, {
-                  decimals: pooledTokenInfo.token.decimals,
-                })}
-              </strong>
-            </div>
-            <span className="dollar-value">{pooledTokenInfo.amountUSD}</span>
-          </li>
-        );})}
+            <li key={index}>
+              <div className="main-info">
+                <MissingLogo
+                  symbol={pooledTokenInfo.token.symbol}
+                  url={pooledTokenInfo.token.logoURI}
+                  width={24}
+                  mobileWidth={24}
+                />
+                <p>
+                  {t("UnstakePosition:overview.unclaimed")}{" "}
+                  {pooledTokenInfo.token.symbol}
+                </p>
+                <strong>
+                  {formatPoolPairAmount(pooledTokenInfo.amount, {
+                    decimals: pooledTokenInfo.token.decimals,
+                  })}
+                </strong>
+              </div>
+              <span className="dollar-value">{pooledTokenInfo.amountUSD}</span>
+            </li>
+          );
+        })}
       </ul>
       <div className="result-section">
         <div className="total-amount-box">
