+++ packages/web/src/hooks/earn/use-remove-position-modal.tsx
@@ -27,7 +27,7 @@ export const useRemovePositionModal = ({
     setModalContent(
       <RemovePositionModalContainer
         allPosition={positions}
-        selectedPosition={selectedPositions}
+        selectedPositions={selectedPositions}
         isWrap={isWrap}
       />,
     );
